TTY input/output and transfer is now possible.
It is now possible to transfer data from the receiver to the transmitter and the other way around. The tty interface isn't used for debugging anymore, but for sending/receiving data.
Next tot that also corrected some errors in the DSM2/DSMX protocol, added a simple SOP, data code and CRC set to the transfer example to test if it works. This all works, but setting the SOP code in the DSM/DSMX protocol isn't working yet.