Changed a little bit the compilation and flashing process in order to make it easyer. Plus documented it a little bit in the README file.
Removed the tests from the default compilation in the makefile. (cause they don't compile by default and there are not mandatory to use the firmware).

Added a file to simplify the flashing procedur.

Removed the PREFIX statement in the flashing makefile.

It's not realy portable, so by default I think it should be considered that the arm toolchain is accessible via the PATH.

Explained a little bit the compilation and flashing process.