[cdcacm] Converted the cdcacm module to use ring buffers.
Converted the cdcacm module to use ring buffers instead of callbacks for
the incoming and outgoing data transfers. We are using a fake process (a
function called periodically from the main loop) to process the outgoing
data and transfer it to the endpoint.

The test works but I can confirm that it is corrupting data. That is
already better than the crashing we were getting before. :)